refactoring names and adding logging if statements
